Move Quoter into test
Eliminates dependency on JUnit in ElixirPsiImplUtil.
